refactor(hackerman): Modulalize the hackerman
